+++ R/compareSpectriPy.R
@@ -9,6 +9,10 @@
 #' [matchms.similarity](https://matchms.readthedocs.io/en/latest/api/matchms.similarity.html).
 #' module.
 #'
+#' **Note**: *SpectriPy* version >= 1.1.3 follows the changes introduced in
+#' *matchms* version 0.32, where the `ModifiedCosine` method was replaced by
+#' `ModifiedCosineGreedy` and `ModifiedCosineHungarian`.
+#'
 #' Selection and configuration of the algorithm can be performed with one of the
 #' *parameter* objects/functions:
 #'
@@ -32,14 +36,22 @@
 #'   (see parameter description for more details). See also
 #'   [matchms CosingHungarian](https://matchms.readthedocs.io/en/latest/api/matchms.similarity.CosineHungarian.html) for more information.
 #'
-#' - `ModifiedCosine()`: The modified cosine score aims at quantifying the
-#'   similarity between two mass spectra. The score is calculated by finding
-#'   the best possible matches between peaks of two spectra. Two peaks are
+#' - `ModifiedCosineGreedy()`: calculate an approximate modified cosine score;
+#'   the modified cosine score aims at quantifying the similarity between two
+#'   mass spectra. The score is calculated by finding the best possible matches
+#'   between peaks of two spectra. This implementation solves the peak
+#'   assignment in a greedy way and is therefore an approximation. Two peaks are
 #'   considered a potential match if their m/z ratios lie within the given
 #'   `tolerance`, or if their m/z ratios lie within the tolerance once a
 #'   mass shift is applied. The mass shift is simply the difference in
 #'   precursor-m/z between the two spectra.
-#'   See also [matchms ModifiedCosine](https://matchms.readthedocs.io/en/latest/api/matchms.similarity.ModifiedCosine.html) for more information.
+#'   See also [matchms ModifiedCosineGreedy](https://matchms.readthedocs.io/en/stable/api/matchms.similarity.html#matchms.similarity.ModifiedCosineGreedy) for more information.
+#'
+#' - `ModifiedCosineHungarian()`: calculate exact modified cosine score between
+#'   mass spectra. The modified cosine score quantifies similarity between two
+#'   mass spectra with optional precursor-based mass shift. The mass shift is
+#'   simply the difference in precursor-m/z between the two spectra.
+#'   See also [matchms ModifiedCosineHungarian](https://matchms.readthedocs.io/en/stable/api/matchms.similarity.html#matchms.similarity.ModifiedCosineHungarian) for more information.
 #'
 #' - `NeutralLossesCosine()`: The neutral losses cosine score aims at
 #'   quantifying the similarity between two mass spectra. The score is
@@ -53,7 +65,6 @@
 #'   on their fingerprints. For this similarity measure to work, fingerprints
 #'   are expected to be derived by running *add_fingerprint()*. See also
 #'   [matchms FingerprintSimilarity](https://matchms.readthedocs.io/en/latest/api/matchms.similarity.FingerprintSimilarity.html) for more information.
-#'
 #'
 #' @note
 #'
@@ -130,15 +141,15 @@
 #' res <- compareSpectriPy(sps, sps[1], param = CosineGreedy())
 #'
 #' ## Calculate pairwise similarity of all spectra in sps with matchms'
-#' ## ModifiedCosine algorithm
-#' res <- compareSpectriPy(sps, param = ModifiedCosine())
+#' ## ModifiedCosineHungarian algorithm
+#' res <- compareSpectriPy(sps, param = ModifiedCosineHungarian())
 #' res
 #'
-#' ## Note that the ModifiedCosine method requires the precursor m/z to be
-#' ## known for all input spectra. Thus, it is advisable to remove spectra
+#' ## Note that the ModifiedCosineHungarian method requires the precursor m/z
+#' ## to be known for all input spectra. Thus, it is advisable to remove spectra
 #' ## without precursor m/z before using this algorithm.
 #' sps <- sps[!is.na(precursorMz(sps))]
-#' compareSpectriPy(sps, param = ModifiedCosine())
+#' compareSpectriPy(sps, param = ModifiedCosineHungarian())
 NULL
 
 setGeneric("compareSpectriPy", function(x, y, param, ...) {
@@ -170,6 +181,12 @@ setClass(
     contains = "CosineGreedy")
 setClass(
     "ModifiedCosine",
+    contains = "CosineGreedy")
+setClass(
+    "ModifiedCosineHungarian",
+    contains = "CosineGreedy")
+setClass(
+    "ModifiedCosineGreedy",
     contains = "CosineGreedy")
 setClass(
     "NeutralLossesCosine",
@@ -221,12 +238,33 @@ CosineHungarian <- function(tolerance = 0.1, mz_power = 0.0,
 #'
 #' @export
 ModifiedCosine <- function(tolerance = 0.1, mz_power = 0.0,
-                                intensity_power = 1.0) {
-    new("ModifiedCosine",
+                           intensity_power = 1.0) {
+    message("'ModifiedCosine' is deprecated! Using 'ModifiedCosineHungarian'")
+    ModifiedCosineHungarian(tolerance, mz_power, intensity_power)
+}
+
+#' @rdname compareSpectriPy
+#'
+#' @export
+ModifiedCosineHungarian <- function(tolerance = 0.1, mz_power = 0.0,
+                                    intensity_power = 1.0) {
+    new("ModifiedCosineHungarian",
         tolerance = as.numeric(tolerance),
         mzPower = as.numeric(mz_power),
         intensityPower = as.numeric(intensity_power)
     )
+}
+
+#' @rdname compareSpectriPy
+#'
+#' @export
+ModifiedCosineGreedy <- function(tolerance = 0.1, mz_power = 0.0,
+                                 intensity_power = 1.0) {
+    new("ModifiedCosineGreedy",
+        tolerance = as.numeric(tolerance),
+        mzPower = as.numeric(mz_power),
+        intensityPower = as.numeric(intensity_power)
+        )
 }
 
 #' @rdname compareSpectriPy
@@ -287,8 +325,17 @@ setMethod("py_fun", "CosineHungarian", function(object) {
                                        object@intensityPower)
 })
 setMethod("py_fun", "ModifiedCosine", function(object) {
-    matchms_similarity$ModifiedCosine(object@tolerance, object@mzPower,
-                                      object@intensityPower)
+    message("'ModifiedCosine' is deprecated; using 'ModifiedCosineHungarian'")
+    matchms_similarity$ModifiedCosineHungarian(object@tolerance, object@mzPower,
+                                               object@intensityPower)
+})
+setMethod("py_fun", "ModifiedCosineHungarian", function(object) {
+    matchms_similarity$ModifiedCosineHungarian(object@tolerance, object@mzPower,
+                                               object@intensityPower)
+})
+setMethod("py_fun", "ModifiedCosineGreedy", function(object) {
+    matchms_similarity$ModifiedCosineGreedy(object@tolerance, object@mzPower,
+                                            object@intensityPower)
 })
 setMethod("py_fun", "NeutralLossesCosine", function(object) {
     matchms_similarity$NeutralLossesCosine(object@tolerance, object@mzPower,